%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/providenceri.R
\name{pvd_arrests}
\alias{pvd_arrests}
\title{Providence police arrests and citations}
\usage{
pvd_arrests(statute_type = NULL, gender = NULL,
  start_date = NULL, end_date = NULL, limit = 1000)
}
\value{
tibble of arrest/citation records
}
\description{
Providence police arrests and citations
}

