%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/providenceri.R
\name{pvd_expenditures}
\alias{pvd_expenditures}
\title{City of Providence expenditures}
\usage{
pvd_expenditures(fiscal_year = NULL, vendor_name = NULL,
  department = NULL, limit = 1000)
}
\value{
tibble of expenditure records
}
\description{
City of Providence expenditures
}

