%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/providenceri.R
\name{pvd_list}
\alias{pvd_list}
\title{List Providence Open Data datasets}
\usage{
pvd_list(limit = 100)
}
\value{
tibble: id, name, category, description, updated_at
}
\description{
List Providence Open Data datasets
}

