%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/providenceri.R
\name{pvd_property_tax}
\alias{pvd_property_tax}
\title{Providence property tax roll}
\usage{
pvd_property_tax(year = 2025, limit = 1000)
}
\value{
tibble of property tax records
}
\description{
Providence property tax roll
}

