%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/providenceri.R
\name{pvd_search}
\alias{pvd_search}
\title{Search Providence Open Data datasets}
\usage{
pvd_search(query, limit = 300)
}
\value{
tibble of matching datasets
}
\description{
Search Providence Open Data datasets
}

