%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/providenceri.R
\name{pvd_view}
\alias{pvd_view}
\title{Query any Providence dataset via SODA}
\usage{
pvd_view(view_id, where = NULL, select = NULL, group = NULL,
  order = NULL, q = NULL, limit = 1000, offset = 0)
}
\value{
tibble of query results
}
\description{
Query any Providence dataset via SODA
}

